requirements.txt
boto3==1.7.28
botocore==1.10.28
Django==2.0.5
django-guardian==1.4.9
django-storages==1.6.6
django-widget-tweaks==1.4.2
docutils==0.14
html5lib==1.0.1
httplib2==0.11.3
jmespath==0.9.3
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.4
reportlab==3.5.2
s3transfer==0.1.13
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.2
